remove set-annotator
