--- internal/pipeline/examples/basicpipeline/.expected/diff.patch
@@ -1,39 +1,0 @@
-diff --git a/.expected/config.yaml b/.expected/config.yaml
-index c0d73f7..6ceab20 100644
---- a/.expected/config.yaml
-+++ b/.expected/config.yaml
-@@ -11,6 +11,8 @@
- # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- # See the License for the specific language governing permissions and
- # limitations under the License.
--
- exitCode: 0
- network: false
-+metadata:
-+  annotations:
-+    builtin/setter-1: 'test'
-diff --git a/configs/resources.yaml b/configs/resources.yaml
-index 0dabe44..0124a57 100644
---- a/configs/resources.yaml
-+++ b/configs/resources.yaml
-@@ -11,11 +11,12 @@
- # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- # See the License for the specific language governing permissions and
- # limitations under the License.
--
- apiVersion: apps/v1
- kind: Deployment
- metadata:
-   name: nginx-deployment
-+  annotations:
-+    builtin/setter-1: 'test'
- spec:
-   replicas: 3
- ---
-@@ -23,3 +24,6 @@ apiVersion: custom.io/v1
- kind: Custom
- spec:
-   image: nginx:1.2.3
-+metadata:
-+  annotations:
-+    builtin/setter-1: 'test'
